+++ .babelrc
@@ -1,0 +1,13 @@
+{
+    "plugins": [
+        ["module:react-native-dotenv", {
+            "moduleName": "@env",
+            "path": ".env",
+            "blocklist": null,
+            "allowlist": null,
+            "safe": false,
+            "allowUndefined": true,
+            "verbose": false
+        }]
+    ]
+}
