+++ .gitignore
@@ -6,3 +6,6 @@
 
 # production
 /build
+
+#environment variables
+.env
