+++ App.js
@@ -1,33 +1,9 @@
-import { StatusBar } from 'expo-status-bar';
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
-import MapView from 'react-native-maps';
+import LoadingPage from './component/LoadingPage';
+
 
 export default function App() {
   return (
-    <View style={styles.container}>
-      <MapView
-        style={styles.map}
-        initialRegion={{
-          latitude: -41.28666552,
-          longitude: 171.772996908,
-          latitudeDelta: 24,
-          longitudeDelta: 12,
-        }}>
-      </MapView>
-    </View>
+    <LoadingPage></LoadingPage>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    ...StyleSheet.absoluteFillObject,
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  map: {
-    ...StyleSheet.absoluteFillObject
-  }
-});
