+++ ios/Podfile
@@ -18,4 +18,6 @@ target 'trackerapptemp' do
   #     flipper_post_install(installer)
   #   end
   # end
+  pod 'react-native-config', :path => '../node_modules/react-native-config'
+
 end
